6th step very tired

+++ src/main/java/com/electric/noinvade/bo/TotalPower.java
@@ -12,5 +12,5 @@ public class TotalPower {
     @Column(name = "time")
     private Instant time;
     @Column(name = "p", tag = true)
-    private Double power;
+    private int power;
 }
